Move project tests to its own directory.
