+++ channels/test/__init__.py
@@ -1,0 +1,2 @@
+from .base import TransactionChannelTestCase, ChannelTestCase, Client, apply_routes  # NOQA isort:skip
+from .http import HttpClient  # NOQA isort:skip
