+++ runtests.py
@@ -7,9 +7,9 @@ from django.conf import settings
 from django.test.utils import get_runner
 
 if __name__ == "__main__":
-    os.environ['DJANGO_SETTINGS_MODULE'] = "channels.tests.settings"
+    os.environ['DJANGO_SETTINGS_MODULE'] = "tests.settings"
     django.setup()
     TestRunner = get_runner(settings)
     test_runner = TestRunner()
-    failures = test_runner.run_tests(["channels.tests"])
+    failures = test_runner.run_tests(["tests"])
     sys.exit(bool(failures))
