+++ tests/test_routing.py
@@ -168,7 +168,7 @@ class RoutingTests(SimpleTestCase):
         Tests inclusion without a prefix
         """
         router = Router([
-            include("channels.tests.test_routing.chatroom_routing"),
+            include("tests.test_routing.chatroom_routing"),
         ])
         self.assertRoute(
             router,
@@ -196,7 +196,7 @@ class RoutingTests(SimpleTestCase):
         Tests route_class with/without prefix
         """
         router = Router([
-            include("channels.tests.test_routing.class_routing"),
+            include("tests.test_routing.class_routing"),
         ])
         self.assertRoute(
             router,
@@ -222,7 +222,7 @@ class RoutingTests(SimpleTestCase):
         Tests inclusion with a prefix
         """
         router = Router([
-            include("channels.tests.test_routing.chatroom_routing", path="^/ws/v(?P<version>[0-9]+)"),
+            include("tests.test_routing.chatroom_routing", path="^/ws/v(?P<version>[0-9]+)"),
         ])
         self.assertRoute(
             router,
@@ -252,7 +252,7 @@ class RoutingTests(SimpleTestCase):
         )
         # Check it works without the ^s too.
         router = Router([
-            include("channels.tests.test_routing.chatroom_routing_nolinestart", path="/ws/v(?P<version>[0-9]+)"),
+            include("tests.test_routing.chatroom_routing_nolinestart", path="/ws/v(?P<version>[0-9]+)"),
         ])
         self.assertRoute(
             router,
@@ -279,7 +279,7 @@ class RoutingTests(SimpleTestCase):
         # Unicode patterns, byte message
         router = Router([
             route("websocket.connect", consumer_1, path="^/foo/"),
-            include("channels.tests.test_routing.chatroom_routing", path="^/ws/v(?P<version>[0-9]+)"),
+            include("tests.test_routing.chatroom_routing", path="^/ws/v(?P<version>[0-9]+)"),
         ])
         self.assertRoute(
             router,
@@ -303,7 +303,7 @@ class RoutingTests(SimpleTestCase):
         # Byte patterns, unicode message
         router = Router([
             route("websocket.connect", consumer_1, path=b"^/foo/"),
-            include("channels.tests.test_routing.chatroom_routing", path=b"^/ws/v(?P<version>[0-9]+)"),
+            include("tests.test_routing.chatroom_routing", path=b"^/ws/v(?P<version>[0-9]+)"),
         ])
         self.assertRoute(
             router,
